MFC 260237: Fix a bug in the HPET emulation where a timer interrupt could be lost when the guest disables the HPET.
The HPET timer interrupt is triggered from the callout handler associated with
the timer. It is possible for the callout handler to be delayed before it gets
a chance to execute. If the guest disables the HPET during this window then the
handler never gets a chance to execute and the timer interrupt is lost.

This is now fixed by injecting a timer interrupt into the guest if the callout
time is detected to be in the past when the HPET is disabled.


git-svn-id: https://svn.freebsd.org/base/stable/10@266477 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f