netinet/in.h: Expose htonl(), htons(), ntohl() and ntohs() in strict POSIX mode.
Put the htonl(), htons(), ntohl() and ntohs() declarations under
__POSIX_VISIBLE >= 200112. POSIX.1-2001 and newer require these to be
exposed from <netinet/in.h> (as well as <arpa/inet.h>).

Note that it may be unnecessary to check __POSIX_VISIBLE >= 200112 because
older versions of POSIX and the C standard do not define this header.
However, other places in the same file already perform the check.

PR:		188316
Submitted by:	Christian Neukirchen


git-svn-id: https://svn.freebsd.org/base/head@266842 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f