Export two new settings for the AM335x PWM, the clock prescaler (clkdiv) and the actual PWM frequency.
Enforce the maximum value for the period sysctl.

The frequency systcl now allows the direct setting of the PWM frequency (it
will try to find the better clkdiv and period for a given frequency, i.e.
the ones that will give the better PWM resolution).

This allows the use lower frequencies on the PWM.  Without changing the
clock prescaler the minimum PWM frequency was 1.52kHz.

PWM frequencies checked with an osciloscope.

PWM output tested with some R/C servos at 50Hz.


git-svn-id: https://svn.freebsd.org/base/head@266937 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f