MFC r263096:
Put the offset of the CRC32C in csum_data instead of 0.
The virtio driver needs the offset to be stored in csum_data,
like in the case for UDP and TCP.

The virtio problem was reported by
Niu Zhixiong <kaiaixi@gmail.com>, who helped in debugging
and testing the patch.


git-svn-id: https://svn.freebsd.org/base/stable/10@267722 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f