MFC	r267490: patch: unsign the line length to avoid overflows.
Patch(1) uses a short int for the line length, which is usually
sufficient for regular diffs, but makes no effort to signal
when there is an overflow.

Change the line length to an unsigned short int to better use
the fact that a length is never negative.  The change is loosely
inspired on a related change in DragonFly, but we avoid spending
more memory than necessary.

While here adjust the messages to be clearer on what is happening.


git-svn-id: https://svn.freebsd.org/base/stable/10@267746 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f