MFC r267755:
Don't take filedesc lock in fdunshare().

We can read refcnt safely and only care if it is equal to 1.

If it could suddenly change from 1 to something bigger the code would be
buggy even in the previous form and transitions from > 1 to 1 are equally racy
and harmless (we copy even though there is no need).


git-svn-id: https://svn.freebsd.org/base/stable/10@268339 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f