Add dl_unwind_find_exidx() for ARM EABI, required for C++ exception handling.  For statically linked apps this uses the __exidx_start/end symbols set up by the linker.  For dynamically linked apps it finds the shared object that contains the given address and returns the location and size of the exidx section in that shared object.
The dl_unwind_find_exidx() name is used by other BSD projects and Android,
and is mentioned in clang 3.5 comments as "the BSD interface" for finding
exidx data.  GCC (in libgcc_s) expects the exact same API and functionality
to be provided by a function named __gnu_Unwind_Find_exidx(), so we provide
that with an alias ("strong reference").

Reviewed by:	kib@
MFC after:	1 week


git-svn-id: https://svn.freebsd.org/base/head@268893 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f